requirements.txt
absl-py @ file:///croot/absl-py_1666362940888/work
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.1.0
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
face-alignment==1.3.5
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.25.0
grpcio==1.14.1
idna==3.4
imageio==2.27.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648562407465/work
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
llvmlite==0.39.1
Markdown @ file:///croot/markdown_1671541909495/work
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib @ file:///croot/matplotlib-suite_1667356714455/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
munkres==1.1.4
networkx==2.6.3
numba==0.56.4
numpy==1.21.6
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging @ file:///croot/packaging_1671697413597/work
Pillow==9.4.0
protobuf==3.20.3
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.2
scikit-image==0.19.3
scipy==1.7.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==2.0.0
tifffile==2021.11.2
torch==1.8.1+cu111
torchaudio==0.8.1
torchsummary==1.5.1
torchvision==0.9.1+cu111
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug @ file:///croot/werkzeug_1671215996852/work
zipp @ file:///croot/zipp_1672387121353/work
